# 7-lecture/Example/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service for store files in base dir",
        "title": "Static File Storage Service (SFSS)",
        "contact": {
            "email": "andrianovartemii@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/delete": {
            "delete": {
                "tags": [
                    "write"
                ],
                "summary": "removes file from base dir by url query param.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "file name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/get/{name}": {
            "get": {
                "tags": [
                    "read"
                ],
                "summary": "eturns file by name from mux.Vars.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "file name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "408": {
                        "description": "Request Timeout",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "tags": [
                    "healt"
                ],
                "summary": "assign ping returns pong (health).",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    }
}